refactor(ModeCircles/Rectangles): Split VehicleMarkers into ModeCircles and ModeRectangles.
